refactor : 회원가입/로그인 로직 리팩토링
