+++ README.md
@@ -14,5 +14,8 @@
   - [x] Product와 Wish : 일대다(1:N) 연관관계 (Product는 여러 Wish 안에 포함될 수 있다)
 
 
-### 2. 테스트해보기 
+### 2. 리팩토링 및 테스트
+- [x] 회원가입/로그인 로직 리팩토링
+- [x] 회원가입/로그인 로직 테스트 
+- [ ] 사용자의 위시리스트에 상품 추가, 수정, 삭제, 조회하는 로직 리팩토링
