+++ src/main/java/gift/Login/common/auth/LoginMember.java
@@ -1,4 +1,4 @@
-package gift.Login.auth;
+package gift.Login.common.auth;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
