+++ src/main/java/gift/Login/common/config/JasyptConfig.java
@@ -1,4 +1,4 @@
-package gift.Login.config;
+package gift.Login.common.config;
 
 import org.jasypt.encryption.StringEncryptor;
 import org.jasypt.encryption.pbe.PooledPBEStringEncryptor;
