+++ src/main/java/gift/Login/common/exception/UserNotFoundException.java
@@ -1,4 +1,4 @@
-package gift.Login.exception;
+package gift.Login.common.exception;
 
 public class UserNotFoundException extends RuntimeException {
     public UserNotFoundException(String message) {
